refactor(ops): split gdn_prefill_fla into ssm_gdn_a3.rs (≤500 LoC cap)
`ssm_gdn_a2.rs` reached 513 LoC — the vtile spine added a kernel handle
and its dispatch arm to `gdn_prefill_fla`, and that function was already
the file's bulk (237 of 513 lines).

Exact piecewise move: `gdn_prefill_fla` and its doc comment go to a
sibling `ssm_gdn_a3.rs` with the same header/imports; `ops.rs` gains the
`#[path]` mod + glob re-export, so every caller keeps reaching it at
`crate::layers::ops::*` unchanged. No behavioural diff.

  ssm_gdn_a2.rs  513 -> 276
  ssm_gdn_a3.rs        253 (new)
